+++ react-ts/src/components/forms/uncontrolled-form/inputs/RadioButton.tsx
@@ -1,13 +1,5 @@
 import { forwardRef } from 'react';
-
-export interface RadioButtonProps {
-  id: string;
-  label: string;
-  type: string;
-  name: string;
-  options: string[];
-  error?: string;
-}
+import { RadioButtonProps } from '../../../../types/interfaces';
 
 const RadioButton = forwardRef<HTMLInputElement[], RadioButtonProps>(
   function RadioButton(props, ref) {
