+++ react-ts/src/types/types.ts
@@ -1,0 +1,5 @@
+export type FieldErrors = {
+  [key: string]: {
+    message: string;
+  };
+};
